@@ -1,6 +1,7 @@
-import React, {useState} from "react";
-import {Button, TextField, IconButton} from "@material-ui/core";
-import {AddBox} from "@material-ui/icons";
+import IconButton from '@mui/material/IconButton/IconButton';
+import TextField from '@mui/material/TextField/TextField';
+import React, {useState} from 'react';
+import {AddBox} from "@mui/icons-material";
 
 //types:
 type AddItemFormPropTypes = {
