@@ -1,5 +1,5 @@
-import React, {useState} from "react";
-import {TextField} from "@material-ui/core";
+import React, {useState} from 'react';
+import {TextField} from "@mui/material";
 
 type EditableSpanPropsType = {
     title: string
